Package everything into respective packages
