+++ src/main/java/CommandsPackage/ByeCase.java
@@ -1,0 +1,13 @@
+package CommandsPackage;
+
+public class ByeCase implements DefaultCase {
+
+    public ByeCase() { }
+
+    @Override
+    public void action() {
+        System.out.println("\nGoodbye! Hope to see you again soon!"+
+                "\n_____________________________________________");
+    }
+
+}
