+++ src/main/java/CommandsPackage/DefaultCase.java
@@ -1,0 +1,7 @@
+package CommandsPackage;
+
+import Exceptions.InvalidCommandException;
+
+public interface DefaultCase {
+    public abstract void action() throws InvalidCommandException;
+}
