+++ src/main/java/Exceptions/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class InvalidCommandException extends Exception {
 
     public InvalidCommandException() {
