+++ src/main/java/TaskPackage/Deadlines.java
@@ -1,0 +1,20 @@
+package TaskPackage;
+
+public class Deadlines extends TasksDefault {
+
+    private String taskType = "[D]";
+    private String deadlineDate;
+
+    public Deadlines(String description, String deadlineDate) {
+        super(description);
+        this.deadlineDate = deadlineDate;
+    }
+
+    @Override
+    public String getDescription() {
+        StringBuilder str = new StringBuilder();
+        str.append(this.taskType).append(super.getDescription()).append(" (by: ").append(deadlineDate).append(")");
+        return str.toString();
+    }
+
+}
