+++ src/main/java/TaskPackage/ToDo.java
@@ -1,0 +1,18 @@
+package TaskPackage;
+
+public class ToDo extends TasksDefault {
+
+    private String taskType = "[T]";
+
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String getDescription() {
+        StringBuilder str = new StringBuilder();
+        str.append(this.taskType).append(super.getDescription());
+        return str.toString();
+    }
+
+}
